consolidate FMSFResource class
